Improved item handling & misc
Better support for items:
*) We now have correct tooltips for armor
*) Armor handling in general is vastly improved
*) Items are now defined via the db (instead of itemTemplates.txt)
*) ALL equipable items now have the correct required level + skills.
*) Slightly improved and cleaner inventory cody

Additionally to the new item stuff, I changed a few minor things:
*) I commented a big part of mapInfo.txt, so the server loads only some
of the maps - which saves a lot of time when debugging ;). I recommend
the Concordia Wilderness map for testing. Feel free to re-add the maps
in case you want to populate them with NPCs/Creatures/Missions.
*) Got rid of some deprecated tables in the database.
*) Added pre-compiled windows binaries

Important: Due to the big changes of the database format, you should
delete all tables and re-import ir_gameserver.sql