Bunch of new stuff and bugfixes
New in this commit:
Predefined AI paths. Yes, paths.
Don't underestimate the use of them. We already had them working in an
earlier version (thanks to dennton), then I destroyed it with the big
combat and AI update. Now they are back, with more functionality.
Paths can be bound to spawnpools, which will make creatures spawned walk
along the path. This can be used to have patroling NPCs or invading
Thrax.
There are three modes for paths: One-Shot (creature will follow path
once, then stop), Cyclic (walk to the end, then start again at
beginning, start and end pathnode must be close together) and Return(go
to end, then follow path in reverse again and repeat).

New commands to setup paths in the game:
.createpath - Starts path creation and immediately creates a path node
on the current player location
.addnode - Adds the next path node to the currently created path on the
current player location
.discardpath - Discard created path without saving
.finishpath - Path is saved to DB, but not used
.spawnpoolpath - Just like .finishpath, but will attach the path to the
LAST created spawnpool (using .addspawn)

If you want to create a spawnpool + a path, do it this way:
Go to where you want to have your creatures spawned - use .addspawnpool
to add the spawn.
Start path creation with .createpath, add nodes with .addnode (which
should be not too far away from the previous node, about 5 ingame
meters)
If you want to have creatures run in a circle, make sure the last
pathnode is near to the first one (less then 5 meters)
Then type .spawnpoolpath to create and link the path. Restart the server
and watch.

The second bigger feature is that armor (the attribute and the items) is
now working correctly. Just give it a try.

Other new features and bugfixes:
*) Added a few example spawns with paths. Check out Alias Das, Thrax
will try to raid it now. :)
*) Some adjustments to damage and healthpoints of the creatures. Still
far away from the original game probably.
*) Readded the turret in Alias Das.
*) The itemtemplate_armor table is now correctly filled with ALL armor
data, now we just have to add an entry to itemtemplate to add a piece of
armor.
*) Added a small "hack" that makes creatures push away each other,
previously it could happen that 5 creatures were standing on a single
spot, which looked ugly.
*) Added basic support for the logos objects (the ones that can be
collected on the map)
*) Recruit lighting ability now working
*) Fixed: Sometimes creature/npc health was displayed wrong
*) Fixed: When creatures attack players, their missiles are no longer
invisible.

As always, new database tables -> reimport ir_gameserver.sql or else the
new version will not work.